+++ app.js
@@ -2,7 +2,7 @@ import express from 'express';
 import logger from 'morgan';
 import cors from 'cors';
 
-import  contactsRouter from './routes/api/contacts';
+import contactsRouter from './routes/api/controllers/contacts/contacts';
 
 const app = express()
