+++ db/contacts.json
@@ -49,7 +49,7 @@
   },
   {
     "id": "9",
-    "name": "Thomas Lucas",
+    "name": "Thomas Lusasa",
     "email": "nec@Nulla.com",
     "phone": "(704) 398-7993"
   },
@@ -72,21 +72,15 @@
     "id": "e78a2524-d6bf-4aac-a49f-0b8db6267f58"
   },
   {
-    "name": "Tiguan Red",
-    "email": "tigred@mail.net",
-    "phone": "(542) 454 456",
-    "id": "e28d4bb8-9ba9-41a9-9f85-823da45392aa"
-  },
-  {
     "name": "ferte Red",
     "email": "tigred@mail.net",
     "phone": "(542) 454 456",
     "id": "970e735f-1f86-4353-8dbd-367fe0f4f7ec"
   },
   {
-    "name": "ferte Red",
+    "name": "ferte Black",
     "email": "tigred@mail.net",
     "phone": "(542) 454 456",
-    "id": "2d16c4d4-6d19-4ad5-972a-67325533e42e"
+    "id": "1c532521-7263-415f-8b95-b18e4dba4e73"
   }
 ]
