+++ models/contacts/getContactsById.js
@@ -1,0 +1,7 @@
+import contacts from '../../db/contacts.json';
+
+
+export const getContactById = async (contactId) => {
+    const [result] = contacts.filter((contact) => contact.id === contactId);
+    return result;
+}
