+++ models/contacts/index.js
@@ -1,0 +1,18 @@
+
+import { listContacts } from './listContacts';
+import { getContactById } from './getContactsById';
+import { addContact} from './addContact';
+import { removeContact } from './removeContact';
+import { updateContact } from './updateContact';
+
+
+export default {
+    listContacts,
+    getContactById,
+    addContact,
+    removeContact,
+    updateContact
+};
+
+
+
