+++ models/contacts/listContacts.js
@@ -1,0 +1,6 @@
+import contacts from '../../db/contacts.json';
+
+
+export const listContacts = async () => {
+    return contacts;
+}
